Automation_Framework/requirements.txt
Flask==3.0.3
pytest==8.2.1
requests==2.32.2
allure-pytest==2.13.2